rename periodic_trigger
