+++ tests/consumers/add_ons/CMakeLists.txt
@@ -2,7 +2,7 @@ set(dir ${CMAKE_CURRENT_SOURCE_DIR})
 
 set(SOURCES
   ${dir}/FilterBlockTest.cpp
-  ${dir}/periodic_trigger.cpp
+  ${dir}/PeriodicTriggerTest.cpp
   ${dir}/RecentRateTest.cpp
   ${dir}/StatusTest.cpp
   ${dir}/ValueFilterTest.cpp
